Mojo: Support lazy serialization.
Implements two new public C API calls: MojoCreateMessage and
MojoGetMessageContext. This constitutes the bulk of what's needed to
support basic lazy message serialization as described in the proposal[1].

Follow-up CLs will introduce additional API surface - namely
MojoSerializeMessage and MojoGetSerializedMessageContents -- and
rename or rework existing APIs to match the details of the proposal.

[1] https://groups.google.com/a/chromium.org/forum/#!topic/chromium-mojo/4CA0Ip56lgA

BUG=725321

+++ mojo/edk/system/handle_table.h
@@ -44,7 +44,7 @@ class HandleTable : public base::trace_event::MemoryDumpProvider {
   // if any of the handles are invalid; or MOJO_RESULT_OK if successful.
   MojoResult BeginTransit(
       const MojoHandle* handles,
-      uint32_t num_handles,
+      size_t num_handles,
       std::vector<Dispatcher::DispatcherInTransit>* dispatchers);
 
   void CompleteTransitAndClose(
